+++ lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:not_too_shabby/view/app.dart';
 
-void main() => runApp(NotTooShabbyApp());
+void main() {
+  runApp(NotTooShabbyApp());
+}